Add extend volumes plugin and disallow reboot in unattend step
Because of an installer patch for the swap disable/enable issue,
the ExtendVolumes plugin has been added by the cloudbase-init
installer to the unattend configuration file.

Also, the cloudbase-init run in the unattend step should not
issue a reboot, as the reboot is performed through the synchronous
command exit code.